refactor: refactor useLocalStorage
