+++ src/hooks/useLocalStorage.ts
@@ -1,22 +1,13 @@
 import { useState, useEffect } from 'react';
+import { getStoredValueFromLS } from './utils/getStoredValueFromLS.ts';
 
 function useLocalStorage<T>(key: string, initialValue: T) {
-  const [storedValue, setStoredValue] = useState<T>(() => {
-    try {
-      const item = localStorage.getItem(key);
-      return item ? JSON.parse(item) : initialValue;
-    } catch (error) {
-      console.error('Error reading from localStorage:', error);
-      return initialValue;
-    }
-  });
+  const [storedValue, setStoredValue] = useState<T>(
+    getStoredValueFromLS(key, initialValue)
+  );
 
   useEffect(() => {
-    try {
-      localStorage.setItem(key, JSON.stringify(storedValue));
-    } catch (error) {
-      console.error('Error writing from localStorage:', error);
-    }
+    localStorage.setItem(key, JSON.stringify(storedValue));
   }, [key, storedValue]);
 
   return [storedValue, setStoredValue] as const;
