+++ src/index.css
@@ -33,8 +33,3 @@ button {
   cursor: pointer;
   transition: border-color 0.25s;
 }
-
-button:focus,
-button:focus-visible {
-  outline: 4px auto -webkit-focus-ring-color;
-}
